@@ -6,32 +6,36 @@
   "nativescript": {
     "id": "br.com.costanonato.taskmanager",
     "tns-android": {
-      "version": "3.2.0"
+      "version": "4.1.3"
     }
   },
   "dependencies": {
-    "@angular/animations": "~4.4.1",
-    "@angular/common": "~4.4.1",
-    "@angular/compiler": "~4.4.1",
-    "@angular/core": "~4.4.1",
-    "@angular/forms": "~4.4.1",
-    "@angular/http": "~4.4.1",
-    "@angular/platform-browser": "~4.4.1",
-    "@angular/router": "~4.4.1",
-    "nativescript-angular": "~4.4.0",
+    "@angular/animations": "~6.0.6",
+    "@angular/common": "~6.0.6",
+    "@angular/compiler": "~6.0.6",
+    "@angular/core": "~6.0.6",
+    "@angular/forms": "~6.0.6",
+    "@angular/http": "~6.0.6",
+    "@angular/platform-browser": "~6.0.6",
+    "@angular/platform-browser-dynamic": "~6.0.6",
+    "@angular/router": "~6.0.6",
+    "nativescript-angular": "^6.0.6",
     "nativescript-social-share": "^1.5.0",
     "nativescript-theme-core": "~1.0.2",
     "reflect-metadata": "~0.1.8",
-    "rxjs": "~5.4.2",
-    "tns-core-modules": "~3.2.0",
-    "zone.js": "~0.8.2"
+    "rxjs": "~6.0.0 || >=6.1.0",
+    "tns-core-modules": "^4.1.1",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
+    "@angular/compiler-cli": "~6.1.0-beta.3",
+    "@ngtools/webpack": "6.1.0-rc.0",
     "babel-traverse": "6.4.5",
     "babel-types": "6.4.5",
     "babylon": "6.4.5",
     "lazy": "1.0.11",
-    "nativescript-dev-typescript": "~0.5.0",
-    "typescript": "~2.4.2"
+    "nativescript-dev-typescript": "^0.7.2",
+    "nativescript-dev-webpack": "^0.14.4",
+    "typescript": "~2.7.2"
   }
 }
