@@ -1,7 +1,8 @@
 import { Response } from "@angular/http";
 import { Injectable } from "@angular/core";
 
-import { Observable } from "rxjs/Observable";
+import { Observable, throwError } from "rxjs";
+import { catchError, map } from "rxjs/operators";
 
 import { TokenService } from "../../../shared/token.service";
 import { Task } from "./task.model";
@@ -18,27 +19,30 @@ export class TaskService{
   public getAll(): Observable<Task[]>{
     let url = `${this.tasksUrl}?q[s]=updated_at+DESC`;
     
-    return this.tokenHttp.get(url)
-      .catch(this.handleErrors)
-      .map((response: Response) => this.responseToTasks(response));
+    return this.tokenHttp.get(url).pipe(
+      catchError(this.handleErrors),
+      map((response: Response) => this.responseToTasks(response))
+    )
   }
 
   
   public getImportant(): Observable<Task[]>{
     let url = `${this.tasksUrl}?q[s]=deadline+ASC`;
     
-    return this.tokenHttp.get(url)
-      .catch(this.handleErrors)
-      .map((response: Response) => this.responseToTasks(response));
+    return this.tokenHttp.get(url).pipe(
+      catchError(this.handleErrors),
+      map((response: Response) => this.responseToTasks(response))
+    )
   }
 
 
   public getById(id: number): Observable<Task> {
     let url = `${this.tasksUrl}/${id}`;
 
-    return this.tokenHttp.get(url)
-      .catch(this.handleErrors)
-      .map((response: Response) => this.responseToTask(response));
+    return this.tokenHttp.get(url).pipe(
+      catchError(this.handleErrors),
+      map((response: Response) => this.responseToTask(response))
+    )
   }
 
 
@@ -46,9 +50,10 @@ export class TaskService{
     let url = this.tasksUrl;
     let body = JSON.stringify(task);
 
-    return this.tokenHttp.post(url, body)
-      .catch(this.handleErrors)
-      .map((response: Response) => this.responseToTask(response));
+    return this.tokenHttp.post(url, body).pipe(
+      catchError(this.handleErrors),
+      map((response: Response) => this.responseToTask(response))
+    )
   }
 
 
@@ -56,33 +61,36 @@ export class TaskService{
     let url = `${this.tasksUrl}/${task.id}`;
     let body = JSON.stringify(task);
 
-    return this.tokenHttp.put(url, body)
-      .catch(this.handleErrors)
-      .map(() => task)
+    return this.tokenHttp.put(url, body).pipe(
+      catchError(this.handleErrors),
+      map((response: Response) => this.responseToTask(response))
+    )
   }
 
 
   public delete(id: number): Observable<null> {
     let url = `${this.tasksUrl}/${id}`;
 
-    return this.tokenHttp.delete(url)
-      .catch(this.handleErrors)
-      .map(() => null)
+    return this.tokenHttp.delete(url).pipe(
+      catchError(this.handleErrors),
+      map(() => null)
+    )
   }
 
 
   public searchByTitle(term: string): Observable<Task[]> {
     let url = `${this.tasksUrl}?q[title_cont]=${term}`;
 
-    return this.tokenHttp.get(url)
-      .catch(this.handleErrors)
-      .map((response: Response) => this.responseToTasks(response));
+    return this.tokenHttp.get(url).pipe(
+      catchError(this.handleErrors),
+      map((response: Response) => this.responseToTasks(response))
+    )
   }
 
 
   private handleErrors(error: Response){
     console.log("SALVANDO O ERRO NUM ARQUIVO DE LOG - DETALHES DO ERRO => ", error);
-    return Observable.throw(error);
+    return throwError(error);
   }
 
 
