@@ -1,6 +1,8 @@
 import { Component, OnInit, ViewContainerRef } from '@angular/core';
-import { ActivatedRoute, Params } from "@angular/router";
+import { ActivatedRoute, ParamMap } from "@angular/router";
 import { FormGroup, FormControl, FormBuilder, Validators } from "@angular/forms";
+
+import { switchMap } from "rxjs/operators";
 
 import { ModalDialogService, ModalDialogOptions } from "nativescript-angular/modal-dialog";
 
@@ -38,13 +40,14 @@ export class TaskDetailComponent implements OnInit{
 
   public ngOnInit(){
     this.task = new Task(null, null);
-
-    this.route.params
-      .switchMap((params: Params) => this.taskService.getById(+params['id']))
-      .subscribe(
-        task => this.setTask(task),
-        error => alert("Ocorreu um no servidor, tente mais tarde.")
-      )
+    
+    this.route.paramMap.pipe(
+      switchMap((params: ParamMap) => this.taskService.getById(+params.get('id')))
+    )
+    .subscribe(
+      task => this.setTask(task),
+      error => alert("Ocorreu um no servidor, tente mais tarde.")
+    )
   }
 
 
